refactor(ProductThumbnail): move it inside the ReviewModal
