--- packages/components/src/molecules/ProductThumbnail/ProductThumbnail.tsx
@@ -1,30 +1,0 @@
-import React, { forwardRef } from 'react'
-import type { HTMLAttributes } from 'react'
-
-export interface ProductThumbnailProps extends HTMLAttributes<HTMLDivElement> {
-  /**
-   * ID to find this component in testing tools (e.g.: cypress,
-   * testing-library, and jest).
-   */
-  testId?: string
-}
-
-const ProductThumbnail = forwardRef<HTMLDivElement, ProductThumbnailProps>(
-  function ProductThumbnail(
-    { testId = 'fs-product-thumbnail', children, ...otherProps },
-    ref
-  ) {
-    return (
-      <div
-        ref={ref}
-        data-testid={testId}
-        data-fs-product-thumbnail
-        {...otherProps}
-      >
-        {children}
-      </div>
-    )
-  }
-)
-
-export default ProductThumbnail
