--- packages/components/src/molecules/ProductThumbnail/index.tsx
@@ -1,8 +1,0 @@
-export { default } from './ProductThumbnail'
-export type { ProductThumbnailProps } from './ProductThumbnail'
-
-export { default as ProductThumbnailImage } from './ProductThumbnailImage'
-export type { ProductThumbnailImageProps } from './ProductThumbnailImage'
-
-export { default as ProductThumbnailTitle } from './ProductThumbnailTitle'
-export type { ProductThumbnailTitleProps } from './ProductThumbnailTitle'
