+++ packages/core/src/components/reviews/ReviewModal/ProductThumbnail/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ProductThumbnail'
+export type { ProductThumbnailProps } from './ProductThumbnail'
