+++ packages/ui/src/styles/components.scss
@@ -35,7 +35,6 @@
 @import "../components/molecules/Modal/styles";
 @import "../components/molecules/NavbarLinks/styles";
 @import "../components/molecules/OrderSummary/styles";
-@import "../components/molecules/ProductThumbnail/styles";
 @import "../components/molecules/ProductCard/styles";
 @import "../components/molecules/ProductCardSkeleton/styles";
 @import "../components/molecules/ProductPrice/styles";
